Dump all the SANITY_CHECK_THIS_CODE sections, updated IMPLEMENT_ME.
This was mostly 1.3 compat code that cared a lot about the internals of
SDL as it stood at the time. Taking a different approach: such as
automatically moving software-rendered apps to the 2.0 Render API, is
probably easier, cleaner, and all around better than trying to coerce the
legacy code to work.