Call SDL_StopTextInput() during init.
Otherwise, macOS might pop up a choose-an-accented-letter window if you hold
down a key, which looks really weird in an FPS.

I had never seen this happen before, so it's not clear why this has to be
done here or if it was a fluke or something. But for now it works, and it's
not like SDL 1.2 has proper text input anyhow.