Always let SDL2 events propagate through the SDL2 event queue.
Previously, our event filter would take what it wanted for the 1.2 app,
and throw _everything_ away. But there are internal event watchers in SDL2
that need to be able to see some events, and the filter prevents that.

In this case: the render API needed to see window resize events to know to
adjust logical scaling, or SDL_WM_ToggleFullscreen() would seriously fail.

Now the events go all the way through the SDL2 queue--never filtered--even
though we still build the 1.2 queue in an event filter. When the user calls
SDL_PumpEvents or SDL_PollEvents, we drain the now-seen-by-everyone-that-cares
events from the SDL2 queue without further processing.

Fixes #25.