Support optional logical scaling for OpenGL rendering.
This lets you do FULLSCREEN_DESKTOP windows with OpenGL games. The game
will render to an FBO and blit it, scaled and centered, to the window
framebuffer, which both avoids the problems on non-_DESKTOP fullscreen
_and_ makes these games work like SDL 1.2 did: find a reasonable resolution
and center the game window in it.

This only works on games that don't use OpenGL framebuffer objects
(for now), so it's disabled unless you set an environment variable
(fow now).