cdrom: Initial shot at fake CD-ROM support.
This makes a directory of MP3 files look like an audio CD to the 1.2 app,
so something like (for example) Quake 1 can rip the retail disc's audio
to the filesystem, and an SDL 1.2 version of the game will still believe the
soundtrack is playing from disc.

This requires us to share the audio device with the 1.2 SDL_OpenAudio
interface, which presents some other tapdancing, but it appears to be a
feasible approach.

This is work in progress; it compiles but there are small pieces missing
and it hasn't been debugged at all. Do not merge yet!