Make fullscreen vidmodes always grab the mouse.
This solves the problem in UT2004 where mouse input was not able to spin all
the way around in fullscreen mode (even though it worked in windowed mode!).
UT2004 explicitly grabbed the mouse for windowed mode, but didn't touch it
for fullscreen, so now we grab by default in fullscreen modes, and don't let
the grab turn off during fullscreen modes in SDL_WM_GrabInput().

Fixes #50.