Handle SDL_WM_GrabInput() when VideoSurface12 is NULL
It's possible for HandleInputGrab() — which is called by
SDL_WM_GrabInput() — to dereference a NULL pointer when attempting to
check if the window is fullscreen. Assume that, if VideoSurface12 is
NULL, we don't have a fullscreen window, and so shouldn't automatically
grab.

This was causing Psychonauts to crash on startup, as it called
SDL_WM_GrabInput() before SDL_SetVideoMode().