Support Multisampling with OpenGL Logical Scaling
OpenGL Logical Scaling didn't work when Multisampling was enabled, as
the faux-backbuffer didn't have multiple samples, so the format didn't
match, making the blit fail and resulting in a black screen.

This change intercepts calls to set up MSAA in SDL_GL_SetAttribute(),
and, if OpenGL Logical Scaling is requested, creates an FBO with the
desired number of samples, and a second FBO used for multisample resolve
before scaling. (It's not possible for glBlitFramebuffer() to scale and
resolve at the same time -- at least without the
GL_EXT_framebuffer_multisample_blit_scaled extension.)

If OpenGL Logical Scaling has not been requested, we quickly
SDL_GL_SetAttribute() the appropriate values for real before creating
the window. This implementation does have the bug that MSAA won't work
if OpenGL Logical Scaling is requested, but fails to initialise, as
we've already created the window without it by the time we know we need
to fall back.