Bind a multisample-resolved READ FBO for OpenGL Scaling
It's possible to glReadPixels() from the default framebuffer if it's
multisampled, but not from an FBO. Work around this by binding last
frame's resolved framebuffer to GL_READ_FRAMEBUFFER if multisampling is
enabled. This could result in bugs from the framebuffer being old, but
nevertheless seems to be an improvement over the existing implementation
(and having to do a resolve on every BindFramebuffer() could potentially
be slow, I guess, though I haven't tested it).