Resolve MSAA framebuffers when default bound/used
It's not possible to read from a multisampled FBO with glReadPixels()
and similar functions unless it is the default framebuffer. Since we're
replacing the default framebuffer, we need some way of triggering an
MSAA resolve before any such read can occur.

To do this, we implement shims for (most of) the functions which can
read from a framebuffer, and — if the default framebuffer is bound as
the READ framebuffer — perform a multisample resolve before running the
orginial function. This ensures that the latest writes to the
faux-backbuffer are reflected in the resolved buffer, which is required
for some games (e.g. Cogs).

In order to implement this, we have glBindFramebuffer's shim cache a
copy of the currently bound Read/Draw framebuffers, which prevents us
from needing to repeatly read it with
glGetIntegerv(GL_DRAW_FRAMEBUFFER_BINDING, …).

For any functions we've missed, we also do a resolve whenever the
default framebuffer is bound. This won't work if there's rendering work
done into that buffer between it being bound and read from, but usually
works pretty well for titles which would read from the framebuffer at
the beginning of a frame.

And, of course, none of this works for the functions defined in
EXT_direct_state_access, but it seems unlikely that there are many games
using that which also use SDL 1.2, so we don't bother for now.