Added replacement 1.2 headers under the zlib license.
These are intentionally cut down to the bare minimum, both to discourage their
use as anything other than a compatibility layer, and because it seems silly
to try to document SDL 1.2's API from scratch at this point.

These headers make an attempt to exist without the need for a configure step;
SDL_config.h is intended to be generic; as such, it makes some assumptions
about the sort of systems that can run SDL2. This might be imperfect and need
patches, and it might not work out in general, but I'm hoping it will.

These files werecreated from a collection of sources: some from-scratch, some
from zlib-licensed SDL2, some from external sources (the 1.2 SDL_opengl.h is
largely under SGI's license), some from lines in 1.2 that I'm confident Sam
or I wrote, and some are "it must be exactly this string of characters and
the Supreme Court just ruled interface copying is Fair Use, shrug."

Fixes #34.