Use scaled mouse coordinates from events in SDL_GetMouseState()
Currently, we pass through SDL_GetMouseState() calls to SDL2 (when not
in relative mouse mode), but SDL20_GetMouseState() returns the actual
coordinates, not the scaled coordinates after
SDL_RenderSetLogicalSize()'s mouse scaling event watch callback is
called.

Keep a track of the mouse coordinates from the events we get (like we do
with relative mouse mode), and return that in SDL_GetMouseState().

(At the moment, we still get the button state from
SDL20_GetMouseState(), which is inconsistent, but not actively a
problem.)