Scale mouse events with OpenGL Logical Scaling
For non-OpenGL windows, mouse coordinates are scaled by
SDL_RenderSetLogicalSize(), but with OpenGL Logical Scaling, we need to
do this ourselves.

Extract the viewport calculations for OpenGL Logical Scaling into their
own function, which is called on mouse events and used to adjust the
mouse coordinates. Because they're adjusted early enough in the event
processing loop, they're already scaled by the time MousePosition is
set.

Note that the scaled events are passed down to any other SDL2 event
handlers that might exist, as we modify the SDL2 events in the filter.
Since this basically matches what SDL_RenderSetLogicalSize() does, it
hopefully won't be a problem.