Don't mark the screen surface as SDL_HWSURFACE.
I'm not sure what this was meant to fix, based on the related comment, but
it causes a problem with the icculus.org port of Build Engine's editor,
which will lock the screen surface because SDL_MUSTLOCK sees SDL_HWSURFACE
and (a bug!) won't unlock it again, causing disaster.

In sdl12-compat, the screen surface is _never_ actually a HWSURFACE; it's
always a buffer of pixels in RAM, not lockable VRAM memory; how we get it to
the screen after that is just an implementation detail, the same as it was
in 1.2 when it might have gone through an X11 server or through win32 GDI.

But we might have to adjust plans here if that SDL_HWSURFACE flag turns out
to have had a specific app in mind to fix, though.

Fixes #71.