Try to match the real SDL 1.2's SDL_stdinc.h better.
Now we try to `#define SDL_fn fn` instead of provide a function declaration
for common things. sdl12-compat always provides the symbols, but this lets
apps compiled with these headers work with an actual build of SDL 1.2 that
might not, because it expected the app to use those macros to reference the
C runtime directly.

This is imperfect because we are trying to do this without a configure stage
for these headers, but it probably already covers most of what we care about
well...but we'll definitely need some #ifdef tweaks for things that aren't
a standardish desktop Linux system...but it's also possible we care less
about binary compatibility with real SDL 1.2 on those targets.

Fixes #73.
(more or less.)