Implemented SDL_OPENGLBLIT.
This works a little differently than SDL 1.2, because we aren't trying to
run this on a Voodoo 2 anymore.

Instead of making a 256x256 texture and uploading/drawing in pieces to
accomodate the small texture, we just make one texture the size of the window.

It's easier and clearer (and I suspect probably faster than the upload/draw
loop that 1.2 used).

This also means your GL needs NPOT textures. Any GPU from the last 15 years
should support this (either through an extension or being OpenGL >= 2.0). We
don't bother trying to deal GL implementation without.

This has only been tested with testgl -logoblit, because, as you should know:
https://discourse.libsdl.org/t/for-all-who-search-a-sdl-openglblit-example/5376/2