Add some common video modes to the modes list
Some modern platforms (e.g. Wayland on KDE) don't expose any video modes
other than the current desktop resolution. For games which expect a
number of common VGA video modes to be available, this can result in
there either being no way to change resolution, or the game failing to
start due to a mode not being available.

Therefore, fake support for a number of common resolutions:
- 640×480
- 800×600
- 1024×768
- 1280×720
- 1920×1080

These are added to the mode list for all available pixel formats for
which there exists a valid, larger mode. The modes are included sorted
in the correct position within the mode list.

The SDL_VideoModeOK() function now uses the fake 1.2 modelist, so picks
up these modes.

If OpenGL Logical Scaling is explicitly disabled by the user, these
modes will not be added, as they may not function correctly if selected
(and the application uses OpenGL).