Add rudimentary emulation of SDL 1.2 scancodes
Introduce a new Scancode20to12() function to convert an SDL 2.0 scancode
into an SDL 1.2 scancode. The actual conversion is platform specific,
because the SDL 1.2 scancodes vary wildly between different platforms
(and, potentially, different configuations or hardware on the same
platform).

The CASESCANCODE20TO12() macro has an idealised 'Linux' scancode (which
is transformed to get the windows scancode), and a separate Mac OS X
scancode. Some scancodes on Linux are deliberately missing (set to 0),
as they differ between Linux configurations, and applications (*cough
DOSBox *cough*) query the system config to work out the values for those
keys. Using 0 triggers a fallback to the keysyms, which work fine.

This is horribly ugly, but that appropriately reflects the horrible
ugliness of how SDL 1.2 scancodes are used.