testloadso: Actually return an error if one occurs
The 'retval' value used to track what exit code the test app should
return is never actually used. Instead 0 (success) is returned
unconditionally, which can cause the program to appear as though it
succeded (at least to a script which doesn't check the text output).

This was caught by an unusued variable warning in gcc, and appears to be
an actual bug, rather than just a useless variable.