Scancode20to12: Fix missing scancode for 'U'
Whoops, this was a copy-paste error on my part: the scancode for 'U' was
missing (set to 0x00) for non Mac OS X systems. On keyboard layouts
which have 'U' in the same position as US English, the fallback to SDLK_
values hid this issue.

The 'U' key should work properly now in things like DOSBox which use
scancodes.