Convert SDL2 Scancodes into SDLKey, not SDL2 Keycodes
SDL 1.2's keysym appears to be entirely layout-indpendent: it always
acts as though there is a US keyboard layout. i.e., the values represent
the positions on the keyboard, not the "meaning" of the key. This
matches what SDL 2.0 scancodes do, but not SDL 2.0 keycodes, which are
translated according to the current keyboard layout.

This means that many SDL 1.2 applications would suddenly start using the
user's keyboard layout, rather than the US one, which -- while quite
convenient in some cases -- is a change in behaviour and could break
things.

Instead of converting SDL_Keycode→SDLKey, this patch converts from
SDL_Scancode→SDLKey, preserving the original behaviour. SDL 1.2
applications will still get translated values in the 'unicode' Keysym
variable if enabled, so text input in many applications will still work
as expected.