The APPLICATION key should actually map to MENU.
This is described in the SDL 2.0 docs as being either a compose key
(notably an actual, hardware compose key like the type seen on old Sun
keyboards) or the windows contextual menu. On my Linux/X11 machine with
a Microsoft USB keyboard, this is the scancode the contextual menu uses.
(What, if anything, ever produces the MENU scancode is a mystery for
another day, I guess...)