checkkeys: Unconditionally display the scancode
It's useful for debugging what scancodes are used by SDL 1.2
(particularly across different platforms) if checkkeys displays it even
when it does successfully map it to an SDLKey.

In addition, print the scancode in hex as well as decimal, as the
hardcoded values we're using in sdl12-compat are all in hex.