Close audio device in SDL_Quit{,Subsystem}
Given we allocate and use a bunch of wrappers around the SDL audio
device, we should shut it down explicitly on quit.

The simplest way to implement this was to just call our SDL_CloseAudio()
implementation in SDL_QuitSubsystem(), which require making it robust
against closing an already closed audio device.

This fixes a sporadic null pointer access in the audio callback on
DOSBox on one of my machines.