Try to deal with surface creations with wonky color masks.
This happens to fix Bug #46 by virtue of how this is implemented, and also
causes an intentional bug in an SDL_perl unit test to pass as expected...
it corrects _bug_ compatibility for specific known issues. It's likely not a
perfect solution, but it's only intended to work around apps with bugs that
SDL 1.2 tolerated.