SDL_syswm.h only includes X11 headers if __has_include can find it.
This isn't a heavily-used feature of SDL 1.2 in general, but we can't promise
X11 is available in a generic sense (we can check for common Windows defines
before including windows.h, but common Unix platforms may or may not have Xlib
available).

In the name of avoiding a configure stage for the headers, we're saying this
only works if we're on a Unix platform (even if X11 is available elsewhere!)
and if the compiler has __has_include support (GCC 5+, Clang, etc).

Note that sdl12-compat itself doesn't currently pass on X11 syswm events, etc,
so it might just be worth dumping this completely and calling apps that rely
on it unsupported...it's entirely possible a given copy of the headers
wouldn't support it in real 1.2, too.