Don't set {R,G,B,A}masks on SDL2 paletted surfaces
SDL2 expects 8-bit surfaces to have correct {R,G,B,A}masks in their
format: either all 0 for paletted surfaces, or the correct values for
SDL_PIXELFORMAT_RGB332. Indeed, SDL_ConvertSurface() will crash if this
is not the case.

Some SDL 1.2 applications provide incorrect masks for paletted surfaces.
We originally zeroed these out in 46fdbf7c1d, but 3af751b91b actually
re-sets these in SetPalette12ForMasks().

Instead, do not set the masks on the SDL 2 surface in
SetPalette12ForMasks(), and force them to be 0 in PixelFormat12to20() if
there is a palette. The masks are still set on the SDL 1.2 surface's
format, and the palette generated by SetPalette12ForMasks() should in
theory make sure surfaces with 8-bit masks still work.

This fixes a crash on startup in circuslinux:
http://www.newbreedsoftware.com/circus-linux/