Fallback to different-bpp modes more easily
A lot of games expect to be able to set the video mode to modes with a
variety of pixel formats, but X11 and Wayland do not permit the system
pixel format to be changed at runtime. Since SDL is able to convert
between these different formats, SDL 1.2 emulated a number of common
formats where the system bpp was >= the desired bpp.

For sdl12-compat, if no modes are available in the desired bpp:
- Expose the mode list from a higher bpp, in the knowledge it can be
  converted.
- Expose 24bpp modes as 32bpp modes if requested, as these are
  functionally identical.

This change mostly updates SDL_ListModes() to fallback to these mode
lists when necessary. It also lets SDL_VideoModeOK() find 24bpp modes if
a 32bpp mode is expected.

With these changes, a number of games which would error out trying to
set 16 and 32 bit modes on an X11/Wayland setup which exposes only
24-bit modes now work with all three bit depths.