Mouse position should be consistant in relative mode
When the mouse is in Relative Mode (MouseInputIsRelative is true), we
"fake" the mouse position by integrating the relative coordinates
(adding them to the MousePosition variable, which accumulates the
position over time).

However, only SDL_GetMouseState() would actually use these
coordinates, events still used the raw position from the system
(possibly scaled, as both the SDL_RenderSetLogicalSize() and OpenGL
based scaling only affects the position, not the relative values). This
creates a mismatch: mouse coordinates received from events don't match
those received from SDL_GetMouseState().

Instead, we should update the mouse coordinates of the SDL 1.2 events
produced to match what we're keeing in the MousePosition variable. We
also do this for mouse button events, as otherwise clicks would hit a
different target from the visible cursor if a software cursor is
rendered.

This is still not perfect: the real and fake mouse positions are
different, so toggling relative mode can result in a "jump" in the mouse
coordinates. Relative mouse movements are also not scaled, so the mouse
movement speed can be different in relative mode.

Tested this against Ren'py SDL 1.2's software mouse cursor option, and
verified no regressions in Wolf4SDL and Multiwinia.