Fix some incorrect scancodes on macOS
The A, N, M, and Keypad + and * keys had the wrong scancodes for OS X. A
was due to a copy-paste error on my part, whereas N and M were also due
to a copy-paste error, but this time in a comment in the DOSBox code.
The Keypad + and * keys were an off-by-one in the table.

The angle bracket key <> on German QWERTZ layouts still seems to be
broken, and will be fixed in a future commit.