Add the NONUSBACKSLASH scancode
Currently the NONUSBACKSLASH key doesn't map to a scancode at all.
It appears to map to 0x32 on macOS[1], and as best I can tell, it is
KEY_102ND in Linux/evdev, which is 0x56 (and since we need to add 8,
should be 0x5E).

DOSBox seems to handle this key specially on Windows (with the 0x56
code, which seems right), and treats it as BACKTICK on macOS. It doesn't
seem to be handled at all on Linux (where it might actually be mapped to
the right-arrow key if evdev detection fails, so this could introduce a
bug here if the SDL_GetWMInfo bits don't work out.)

[1] https://github.com/libsdl-org/sdl12-compat/issues/84#issuecomment-882555428