Scale relative mouse motion with OpenGL Logical Scaling
This makes it match the SDL_Render-based scaling, which does this.

The SDL_MOUSE_RELATIVE_SCALING hint is also now supported for OpenGL
scaling, again to match the SDL_Render path. It's enabled by default,
though obviously is a no-op if OpenGL scaling is not used.

Note that this _doesn't quite_ match the SDL_Render version, in that
it's not accumulating the remainer. This'll be fixed in a follow-up
commit.