Do not generate VIDEORESIZE events on SIZE_CHANGED
The recommended way to handle VIDEORESIZE is with a new
SDL_SetVideoMode() call, and creating a new window generates a
SIZE_CHANGED event, which leads to an infinite loop. The other main case
for SIZE_CHANGED events is when the screen resolution is changed out
from underneath a fullscreen window, and we handle that with the logical
scaling.

So it doesn't appear that there's any use for SIZE_CHANGED events in SDL
1.2 applications. So let's just ignore them. This fixes Drawf Fortress
in bug #127.