Add SDL12COMPAT_USE_KEYBOARD_LAYOUT hint
SDL 1.2 seems to only account for (non-en_US) keyboard layouts on some
platforms and configurations. In particular, the KeySym may contain the
translated key, or may contain a "scancode" which is layout-independent.

In particular, on X11 the configured layout is used, but only (it seems)
if it is the only configured layout. (On systems where multiple layouts
are configured, the US layout is used regardless of the configured one.)
Similarly, the windib backend deliberately forces the US layout in order
to match the DirectX backend, which only provides scancodes.

Since this behaviour is very inconsistant between different
configurations, this patch provides a way of configuring which is used
at runtime:

If the SDL12COMPAT_USE_KEYBOARD_LAYOUT environment variable is set to a
nonzero value, sdl12compat will translate keyboard input according to
the current layout. This matches SDL 1.2 on X11 with only one configured
layout. Otherwise (the default), sdl12compat will use SDL2 scancodes for
keyboard input, essentially forcing all input to the US layout. This
matches SDL 1.2 on Windows, and on X11 with multiple configured layouts.

Regardless of the value of this environment variable, the 'unicode'
value will be affected by keyboard layout.

For discussion and more details see:
- sdl12compat issue #135: Non-US Keyboard layouts not supported
  https://github.com/libsdl-org/sdl12-compat/issues/135
- sdl12compat PR #97: Use Scancodes instead of Keycodes
  https://github.com/libsdl-org/sdl12-compat/pull/97
- SDL 1.2 issue #95: windib returns locale dependent SDLK_ codes
  https://github.com/libsdl-org/SDL-1.2/issues/95