Make SDL12COMPAT_USE_KEYBOARD_LAYOUT's default platform-specific.
It looks like (at least out of the top three of Windows, Mac, and
Linux), only Windows forces the use of the US layout. Mac (at least from
a cursory reading of the code), translates according to the keyboard
layout, as does X11 (albeit with some bugs).

In this case, set the default to disable layout translation only on
Win32, and have it enabled by default on other platforms.

This does break DOSBox, but only when "usescancodes = false" (or
potentially if keyboard layouts changed some of the non-printable
character keys which would break out scancode fallback). Since this is
not the default, hopefully people won't encounter it...?