Time presentation of partial screen updates based on refresh rate
Currently, for applications which update only part of the screen, we
present the screen every 15ms. Instead, update based on the refresh
rate. For fullscreen applications, this will be the refresh rate of the
screen they're on, for windowed applications, we use the refresh rate of
the display selected with SDL_VIDEO_FULLSCREEN_DISPLAY.

The calculation is split out into the GetDesiredMillisecondsPerFrame()
function, which is then used for both YUV and UpdateRects() updates.