Make SDL_EnableKeyRepeat() just use the SDL2 repeat field
This isn't a proper implementation which supports the corrent delay and
interval fields, but I still think it's better than nothing.

In particular, while we were trying to suppress key repeat events
before, we still got them because of SDL_TEXTINPUT events, which took
repeat into account, so if repeat is disabled, we just get a bunch of
events with SDLK_UNKNOWN (and a valid unicode char). Now, if key repeat
is enabled, at least those events look sensible, even if the timing is
not exactly what might've been wanted.