SDL_SetVideoMode: do window resize twice to workaround SDL2 (x11?) issue.
It looks like at the moment SDL2/x11 misbehaves if you resize a fullscreen
window with SDL_SetWindowSize and then make it not fullscreen with
SDL_SetWindowFullscreen; it doesn't appear to accept the new window size
when it leaves fullscreen mode. But a second SDL_SetWindowSize() at this
point appears to be accepted.

I believe this to be an SDL2 issue, but the workaround is easy enough, that
we'll keep this for older SDLs even if a newer release fixes the issue.

Fixes #148.