api: Implement X11_KeyToUnicode.
This is for compat with some existing games that touched an internal piece
of SDL; it was not part of the public 1.2 API.

This is an _extremely_ naive implementation of this function, but I'm hoping
it's good enough for the few things that use it. Patches that don't make this
function ridiculous are welcome, though.

Fixes #153.