fixme removal: "wrap surface"
Not sure _what_ this meant, but it was likely saying at one point that we
need to convert the 1.2 surface to be 2.0 compatible...but we obviously do
that now.

It might have just been a note to myself to finish implementing this thing,
and I forgot to remove the FIXME after? I'm not sure.

Reference Issue #143.