Handle SDL12_Surface::refcount in SDL_FreeSurface
SDL_Surface has a refcount field, and surfaces shouldn't be destroyed
until it's zero.

Confusingly, in sdl12-compat, there are two refcount fields: one for the
SDL 1.2 surface, and one for the SDL 2.0 surface used to implement it.

There are two ways this could be implemented:
1) Keep the SDL12_Surface and SDL20_Surface refcounts synchronised
   across every call into SDL2.
2) Completely separate the refcounts.

This implements option 2: the SDL12_Surface's refcount is set to 1 when
created, not to be equal to the SDL2 surface's refcount. We then
decrement this and free it as required.

This is safe for two reasons:
- As long as the 1.2 surface's refcount is > 0, there must be at least
  one reference to the 2.0 surface.
- It turns out nothing in SDL 2 is actually using the SDL 2 surface
  refcount.

This makes "frogatto"[1] not crash on startup with a double-free of
SDL_Surfaces, as it makes heavy use of the SDL 1.2 refcount field[2].

[1]: https://frogatto.com/download/
[2]: https://github.com/frogatto/frogatto/blob/1.2_stable/src/surface.hpp