Implement a "quirks" system for application-specific workarounds
There are enough application-specific compatibility workarounds that it
may be worth keeping a database of them and automatically applying them.

This implementation adds a "quirks table": an array listing the
executable name and a series of string→string maps which match the
"hints" we've previously used manually.

All of those hints now go via an SDL12Compat_GetHint() function, which
will check the environment variable first, then look up a quirk based on
the executable name, then fallback to the default. Apart from the
compatibility table, this roughly reflects the SDL2 Hints system (albeit
simplified somewhat). This includes a new SDL12Compat_GetHintBoolean()
function which does most of the nasty calls to atoi() and
ternary-operator use which previously happened for every environment
variable check.

It also adds a few quirks by default (for Awesomenauts, Braid, and
DOSBox), and will print out any quirks associated with the application
on startup if SDL12COMPAT_DEBUG_LOGGING is enabled.