Resolve relative paths when initialising CD audio
Some games will chdir() in between initialising SDL and trying to play
CD audio, which will break if the path passed to
SDL12COMPAT_FAKE_CDROM_PATH is relative. Instead, resolve the relative
path to an absolute path in InitializeCDSubsystem(). It's still going to
be a pain for the user if the game calls chdir() before even SDL is
initialised, but it's better than nothing.

Implemented for (and tested on) Win32 and unix. Anything with realpath()
should work.