quirk: Disable SDL_QuitSubsystem(SDL_INIT_VIDEO) for Multiwinia/steam
For some reason, the Steam build of Multiwinia seems to be missing a
call to SDL_Init() where the (older) non-Steam builds have one. This
results in the game, when a resolution change is requested:
- Calling SDL_QuitSubsystem(SDL_INIT_VIDEO) when the old window is
  destroyed.
- NOT calling SDL_Init(SDL_INIT_VIDEO) when creating the new window.
- Calling SDL_GetVideoInfo(), and promptly asserting when NULL is
  returned.
- (And then proceed to create a window, which will fail, etc…)

This issue _is_ reproducable under real SDL 1.2 (though with
SDL_VIDEO_FULLSCREEN_HEAD set, no other modes are listed, so it's
difficult to trigger), but it seemed worth working around.

To do so, simply make SDL_QuitSubsystem() mask out SDL_INIT_VIDEO. This
is hidden behind a new SDL12COMPAT_NO_QUIT_VIDEO hint, which is enabled
by a quirk for "multiwinia.bin.x86".

The various non-steam builds of the game do have the appropriate call to
SDL_Init(), and the Steam build is 32-bit only, so we only activate this
hack for the 32-bit builds.

Steam App ID: 1530
Steam Build ID: 4497829