FIXME removal: We _could_ request a simpler GL context, but don't.
It's risky to find out we broke something this way when nothing is
actually having a problem at the moment, we'd have to juggle the state
between what SDL 1.2 is pretending for the app and what SDL2 actually
used, and we're not talking about _that_ large a resource savings anyhow.

Reference Issue #143.