video: If drawing the same overlay twice in a frame, SDL_Flip() first.
This will update the screen with the current surface contents and the
existing overlay on top of it.

This is sort of a hack, but probably closer to how this would have to
work in real life for this corner case, which you can see in
test/testoverlay.c ... in practice, basically everything that does this
does so with an event loop, and code for animation timing--not to mention
using a single overlay per frame, so this is really getting into the
pathological what-ifs weeds, mostly.

But now testoverlay.c works and nothing else should break, I think.

Fixes #193.