video: Always process and save off the surface from SDL_WM_SetIcon().
Previously, if there was already a window, we'd just pass it through to
SDL2's SDL_SetWindowIcon(), but:

- This way if we destroy the window for various reasons (like recreating it
because SDL_SetVideoMode is changing the format, etc), we'll be able to
reset the icon, which we couldn't before.

- Before, there were two paths: one where we process the icon before handing
it to SDL2 and one where we didn't. Notably, before, existing windows would
ignore the mask passed to SDL_WM_SetIcon.

This was probably a lot of explanation for what was an accidental line of
code instead of an intentional but buggy design decision.  :)

Fixes #191.