FIXME removal: Give up on SDL 1.2 SDLK_(COMPOSE|BREAK|EURO) keycodes.
SDLK_COMPOSE was only referenced in the x11 backend, as a sort of "I have
nothing better to assign to this random set of keys." SDLK_BREAK is generally
the same key as SDLK_PAUSE. SDLK_EURO appears to only be used by the 1.2
BeOS backend, and The Be Book can't identify a physical key that generates
this keycode:

https://www.haiku-os.org/legacy-docs/bebook/TheKeyboard_KeyCodes.html

So nothing reasonable maps to these things (and if they _are_ reasonable,
SDL2 doesn't report anything equivalent to them), so in the best case they
aren't goign concerns, in the worst case we just pretend the user never
pressed them.

So the FIXME is simply removed.

Reference Issue #143.