quirks: Force "tucnak" to x11, software renderer, and no framebuffer texture.
It uses rendering and the event loop from different threads, so it upsets
OpenGL and is hard to work around, beyond just falling back to an
all-software rendering path (which means it has to use X11 too).

There is a real SDL2 target for this app, which doesn't have these
limitations, but this gets the SDL 1.2 target working.

Fixes #182.