video: Try to reset current GL context after touching the renderer.
This way if an app tries to render on a background thread, SDL2's GL
renderers will have a chance of working, since they'll be able to set
the context current for that thread.

This does not fix tucnak, for reasons that aren't clear, so that'll
continue to force the software renderer through the quirks mechanism, but
it definitely seems to fix linapple and probably other things, too.

Fixes #104.
Fixes #110.
Fixes #181.
Fixes #155.

(I think.)