audio: Always open the audio device at at least CD audio quality.
We have to be prepared to convert/buffer between the app's wants and
SDL2's anyhow, and this lets us make sure that if the app tries to play
a CD after opening audio at a low quality, we don't have to reopen the SDL2
audio device to accomodate it.

Fixes #202.