Set DYLIB_COMPAT_VERSION to match SDL-1.2 autotools
The compatibility_version is meant to go up when new symbols are added
to a library, so an executable built against an older version of a
library can run against a newer version of that library, but not the
other way round. The most binary compatible approach is therefore to
use the highest of the different compatibility_version numbers that
could have been used by SDL-1.2.