Clamp mouse coordinates to the backbuffer size for non-OpenGL
When we're logically scaling, and preserving the aspect ratio (so are
letterboxing or pillarboxing), the scaled mouse coordinates may be
outside the application window (backbuffer). These can even be negative,
which get underflowed to very large positive values, as SDL 1.2 uses
unsigned coordinates here.

We were already clamping these values for OpenGL logical scaling, which
we do ourselves, but the event filter used with SDL_Renderer doesn't, so
we'll manually clamp these all the time.