events: Key repeat waits the delay PLUS one interval at start.
In 1.2, this just sets a flag when the delay is done and restarts
the counter for the first interval, but doesn't send a repeat
keydown event at this point. To simplify, we just set the initial
wait to be the delay plus the interval instead of just the delay.

Fixes #251.